requirements.txt
# K-Means Trading Strategy Dependencies
# Core scientific computing and data analysis
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Machine learning
scikit-learn>=1.0.0

# Interactive Brokers API
ibapi>=9.81.1

# WebSocket communication
websockets>=10.0

# Configuration management
PyYAML>=6.0

# Optional: Enhanced logging and monitoring
# coloredlogs>=15.0

# Development dependencies (uncomment for development)
# pytest>=6.0.0
# pytest-cov>=2.12.0
# black>=21.0.0
# flake8>=3.9.0
# mypy>=0.910